requirement.txt
Flask-API
redis
python-dotenv